refactor: isolate XML parsing logic for improved testability
The `xml_parser` was previously implemented as a monolithic file that was difficult to test in isolation. It relied on direct filesystem access for assets and mutated global `params::params`, making unit tests brittle and complex.

This refactoring decouples the parsing logic from its dependencies:

- A new `xml_parser_utils` layer now contains the core XML traversal and object-building logic, which can be tested independently.
- Introduced a `ParserContext` struct to manage state locally, eliminating direct mutation of global parameters during the parsing phase.
- Asset loading for waveforms, antennas, and targets is now handled via a dependency-injected `AssetLoaders` struct, allowing for mocks to be used in tests.

The main `xml_parser.cpp` now acts as a thin facade, responsible only for setting up the real file loaders and committing the parsed `ParserContext` to the global state after a successful parse. This change significantly improves the modularity and testability of the XML parsing subsystem.

Closes #173
