+++ packages/libfers/src/CMakeLists.txt
@@ -45,6 +45,7 @@ set(LIBFERS_PRIVATE_HEADERS
 	${CMAKE_CURRENT_LIST_DIR}/serial/libxml_wrapper.h
 	${CMAKE_CURRENT_LIST_DIR}/serial/waveform_factory.h
 	${CMAKE_CURRENT_LIST_DIR}/serial/xml_parser.h
+	${CMAKE_CURRENT_LIST_DIR}/serial/xml_parser_utils.h
 	${CMAKE_CURRENT_LIST_DIR}/signal/dsp_filters.h
 	${CMAKE_CURRENT_LIST_DIR}/signal/radar_signal.h
 	${CMAKE_CURRENT_LIST_DIR}/timing/prototype_timing.h
@@ -83,6 +84,7 @@ set(LIBFERS_SOURCES
 	${CMAKE_CURRENT_LIST_DIR}/serial/waveform_factory.cpp
 	${CMAKE_CURRENT_LIST_DIR}/serial/response.cpp
 	${CMAKE_CURRENT_LIST_DIR}/serial/xml_parser.cpp
+	${CMAKE_CURRENT_LIST_DIR}/serial/xml_parser_utils.cpp
 	${CMAKE_CURRENT_LIST_DIR}/signal/dsp_filters.cpp
 	${CMAKE_CURRENT_LIST_DIR}/signal/radar_signal.cpp
 	${CMAKE_CURRENT_LIST_DIR}/timing/prototype_timing.cpp
